requirements.txt
# Data Processing
pandas==2.0.0
numpy==1.23.5
scikit-learn==1.2.2

# Web Scraping
beautifulsoup4==4.12.0
requests==2.28.2

# Machine Learning
torch==2.0.0
pymc==5.1.2
arviz==0.15.1
joblib==1.2.0

# Visualization
matplotlib==3.7.1
seaborn==0.12.2

# Cloud Integration
google-cloud-storage==2.8.0
google-cloud-functions==1.13.0

# Dashboard
streamlit==1.22.0

# Deployment
functions-framework==3.3.0
flask==2.2.3

# Testing
pytest==7.3.1